fix: add non-null assertion operator to resource id emissions
The Resource interface defines id as optional (id?: number) because we might import objects from a template or pasting, which don't have an ID. But our emit handlers expect a definite number. Added the non-null assertion operator (!) to tell TypeScript we're certain these values won't be undefined in this context.